requirements.txt
hydra-core
wandb
Pillow
# clip @ git+https://github.com/openai/CLIP.git@beba48f35392a73c6c47ae67ddffced81ad1916d
salesforce-lavis
pytorch-lightning